backend/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
channels==2.4.0
django-filter==2.2.0
djangochannelsrestframework==0.0.5
drf-yasg==1.17.1
pytest==5.4.0
pytest-django==3.9.0
pytest-asyncio==0.11.0
channels_redis==2.4.2
Pillow==7.1.1
openpyxl==3.0.3
bs4
requests
easy-thumbnails
django-image-cropping==1.4.0
django-cors-headers
django-rest-framework-social-oauth2 
social-auth-app-django